Fix types in beat tracking code (#1977)
The development version of Numba 0.62.0 is unhappy about `range`
being applied to floating point values here.

+++ librosa/beat.py
@@ -583,7 +583,7 @@ def __beat_track_dp(localscore, frames_per_beat, tightness, backlink, cumscore):
         # Search over all possible predecessors to find the best preceding beat
         # NOTE: to provide time-varying tempo estimates, we replace
         # frames_per_beat[0] by frames_per_beat[i] in this loop body.
-        for loc in range(i - np.round(frames_per_beat[tv * i] / 2), i - 2 * frames_per_beat[tv * i] - 1, - 1):
+        for loc in range(i - round(frames_per_beat[tv * i] / 2), int(i - 2 * frames_per_beat[tv * i] - 1), - 1):
             # Once we're searching past the start, break out
             if loc < 0:
                 break
